feat(bounded-queue): add support to bound task queues
