+++ .github/workflows/main.yml
@@ -10,7 +10,7 @@ jobs:
     name: Run tests
     strategy:
       matrix:
-        go-version: [1.23.x, 1.22.x, 1.21.x]
+        go-version: [1.23.x, 1.22.x, 1.21.x, 1.20.x]
         os: [ubuntu-latest, macos-latest, windows-latest]
     runs-on: ${{ matrix.os }}
     steps:
