+++ go.mod
@@ -1,5 +1,3 @@
 module github.com/alitto/pond/v2
 
-go 1.21
-
-require golang.org/x/sync v0.10.0
+go 1.20
