--- internal/dispatcher/dispatcher_test.go
@@ -1,178 +1,0 @@
-package dispatcher
-
-import (
-	"context"
-	"sync"
-	"sync/atomic"
-	"testing"
-	"time"
-
-	"github.com/alitto/pond/v2/internal/assert"
-)
-
-func TestDispatcher(t *testing.T) {
-
-	ctx := context.Background()
-
-	writers := 100
-	writeCount := 1000
-
-	writerWg := sync.WaitGroup{}
-	writerWg.Add(writers)
-	readerWg := sync.WaitGroup{}
-	readerWg.Add(writers * writeCount)
-	var receivedCount atomic.Uint64
-
-	receiveFunc := func(elems []int) {
-		for range elems {
-			receivedCount.Add(1)
-			readerWg.Done()
-		}
-	}
-
-	dispatcher := NewDispatcher(ctx, receiveFunc, 1024)
-
-	// Assert counters
-	assert.Equal(t, uint64(0), dispatcher.Len())
-	assert.Equal(t, uint64(0), dispatcher.WriteCount())
-	assert.Equal(t, uint64(0), dispatcher.ReadCount())
-
-	// Launch goroutines that submit many elements to the unbounded channel
-	for i := 0; i < writers; i++ {
-		workerNum := i
-		go func() {
-			for i := 0; i < writeCount; i++ {
-				dispatcher.Write(workerNum*10000 + i)
-			}
-			writerWg.Done()
-		}()
-	}
-
-	// Wait for both readers and writers
-	writerWg.Wait()
-	dispatcher.CloseAndWait()
-	readerWg.Wait()
-
-	// Assert counters
-	assert.Equal(t, uint64(writers*writeCount), receivedCount.Load())
-	assert.Equal(t, uint64(0), dispatcher.Len())
-	assert.Equal(t, uint64(writers*writeCount), dispatcher.ReadCount())
-	assert.Equal(t, uint64(writers*writeCount), dispatcher.WriteCount())
-}
-
-func TestDispatcherWithContextCanceled(t *testing.T) {
-
-	ctx, cancel := context.WithCancel(context.Background())
-
-	receivedCount := atomic.Uint64{}
-	receiveFunc := func(elems []int) {
-		for range elems {
-			receivedCount.Add(1)
-		}
-	}
-
-	dispatcher := NewDispatcher(ctx, receiveFunc, 1024)
-
-	// Assert counters
-	assert.Equal(t, uint64(0), dispatcher.Len())
-	assert.Equal(t, uint64(0), dispatcher.WriteCount())
-	assert.Equal(t, uint64(0), dispatcher.ReadCount())
-
-	// Cancel the context
-	cancel()
-
-	// Attempt to write to the dispatcher
-	err := dispatcher.Write(1)
-
-	assert.Equal(t, ErrDispatcherClosed, err)
-
-	// Assert counters
-	assert.Equal(t, uint64(0), dispatcher.Len())
-	assert.Equal(t, uint64(0), dispatcher.WriteCount())
-	assert.Equal(t, uint64(0), dispatcher.ReadCount())
-}
-
-func TestDispatcherWithContextCanceledAfterWrite(t *testing.T) {
-
-	ctx, cancel := context.WithCancel(context.Background())
-
-	var receivedCount atomic.Uint64
-	receiveFunc := func(elems []int) {
-		for range elems {
-			receivedCount.Add(1)
-		}
-	}
-
-	dispatcher := NewDispatcher(ctx, receiveFunc, 1024)
-
-	// Assert counters
-	assert.Equal(t, uint64(0), dispatcher.Len())
-	assert.Equal(t, uint64(0), dispatcher.WriteCount())
-	assert.Equal(t, uint64(0), dispatcher.ReadCount())
-
-	dispatcher.Write(1)
-	time.Sleep(5 * time.Millisecond)
-
-	// Cancel the context
-	cancel()
-
-	dispatcher.Write(1)
-	time.Sleep(5 * time.Millisecond)
-
-	// Assert counters
-	assert.Equal(t, uint64(0), dispatcher.Len())
-	assert.Equal(t, uint64(1), dispatcher.WriteCount())
-	assert.Equal(t, uint64(1), dispatcher.ReadCount())
-}
-
-func TestDispatcherWriteAfterClose(t *testing.T) {
-
-	ctx := context.Background()
-
-	receivedCount := atomic.Uint64{}
-	receiveFunc := func(elems []int) {
-		for range elems {
-			receivedCount.Add(1)
-		}
-	}
-
-	dispatcher := NewDispatcher(ctx, receiveFunc, 1024)
-
-	// Close the dispatcher
-	dispatcher.Close()
-
-	// Write to the dispatcher
-	err := dispatcher.Write(1)
-	time.Sleep(5 * time.Millisecond)
-
-	// Assert counters
-	assert.Equal(t, ErrDispatcherClosed, err)
-	assert.Equal(t, uint64(0), dispatcher.Len())
-	assert.Equal(t, uint64(0), dispatcher.WriteCount())
-	assert.Equal(t, uint64(0), dispatcher.ReadCount())
-}
-
-func TestDispatcherWriteAfterCloseConcurrent(t *testing.T) {
-
-	ctx := context.Background()
-
-	dispatcher := NewDispatcher(ctx, func(t []int) {}, 1024)
-
-	wg := sync.WaitGroup{}
-	defer wg.Wait()
-
-	for i := 0; i < 1024; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-
-			assert.NotPanics(t, func() {
-				dispatcher.Write(1)
-			})
-		}()
-	}
-
-	assert.NotPanics(t, func() {
-		dispatcher.Close()
-	})
-}
