+++ pool_test.go
@@ -156,7 +156,7 @@ func TestPoolMetrics(t *testing.T) {
 
 func TestPoolSubmitOnStoppedPool(t *testing.T) {
 
-	pool := newPool(100)
+	pool := newPool(100, nil)
 
 	pool.Submit(func() {})
