+++ result.go
@@ -51,17 +51,21 @@ func (p *resultPool[R]) submit(task any) Result[R] {
 
 	wrapped := wrapTask[R, func(R, error)](task, resolve)
 
-	p.dispatcher.Write(wrapped)
+	p.pool.submit(wrapped)
 
 	return future
 }
 
 func (p *resultPool[R]) NewSubpool(maxConcurrency int, options ...Option) ResultPool[R] {
-	return newResultSubpool[R](maxConcurrency, p.Context(), p.pool, options...)
+	return newResultPool[R](maxConcurrency, p.pool, options...)
+}
+
+func newResultPool[R any](maxConcurrency int, parent *pool, options ...Option) *resultPool[R] {
+	return &resultPool[R]{
+		pool: newPool(maxConcurrency, parent, options...),
+	}
 }
 
 func NewResultPool[R any](maxConcurrency int, options ...Option) ResultPool[R] {
-	return &resultPool[R]{
-		pool: newPool(maxConcurrency, options...),
-	}
+	return newResultPool[R](maxConcurrency, nil, options...)
 }
